+++ src/treq/client.py
@@ -2,15 +2,20 @@ from __future__ import absolute_import, division, print_function
 
 import mimetypes
 import uuid
+import warnings
 
-from io import BytesIO
+import io
+
+import six
+from six.moves.collections_abc import Mapping
+from six.moves.http_cookiejar import CookieJar
+from six.moves.urllib.parse import urlencode as _urlencode
 
 from twisted.internet.interfaces import IProtocol
 from twisted.internet.defer import Deferred
 from twisted.python.components import proxyForInterface
-from twisted.python.compat import _PY3, unicode
 from twisted.python.filepath import FilePath
-from hyperlink import URL
+from hyperlink import DecodedURL, EncodedURL
 
 from twisted.web.http_headers import Headers
 from twisted.web.iweb import IBodyProducer, IResponse
@@ -27,31 +32,17 @@ from twisted.web.client import (
 from twisted.python.components import registerAdapter
 from json import dumps as json_dumps
 
-from treq._utils import default_reactor
 from treq.auth import add_auth
 from treq import multipart
 from treq.response import _Response
 from requests.cookies import cookiejar_from_dict, merge_cookies
 
-if _PY3:
-    from urllib.parse import urlencode as _urlencode
 
-    def urlencode(query, doseq):
-        return _urlencode(query, doseq).encode('ascii')
-    from http.cookiejar import CookieJar
-else:
-    from cookielib import CookieJar
-    from urllib import urlencode
+_NOTHING = object()
 
-try:
-    # The old location was quixotically deprecated and might actually be
-    # removed in 3.10, maybe.
-    #
-    # See https://github.com/html5lib/html5lib-python/issues/419 for more of
-    # this tale of woe.
-    from collections.abc import Mapping
-except ImportError:
-    from collections import Mapping
+
+def urlencode(query, doseq):
+    return six.ensure_binary(_urlencode(query, doseq), encoding='ascii')
 
 
 class _BodyBufferingProtocol(proxyForInterface(IProtocol)):
@@ -116,36 +107,42 @@ class HTTPClient(object):
         """
         See :func:`treq.get()`.
         """
+        kwargs.setdefault('_stacklevel', 3)
         return self.request('GET', url, **kwargs)
 
     def put(self, url, data=None, **kwargs):
         """
         See :func:`treq.put()`.
         """
+        kwargs.setdefault('_stacklevel', 3)
         return self.request('PUT', url, data=data, **kwargs)
 
     def patch(self, url, data=None, **kwargs):
         """
         See :func:`treq.patch()`.
         """
+        kwargs.setdefault('_stacklevel', 3)
         return self.request('PATCH', url, data=data, **kwargs)
 
     def post(self, url, data=None, **kwargs):
         """
         See :func:`treq.post()`.
         """
+        kwargs.setdefault('_stacklevel', 3)
         return self.request('POST', url, data=data, **kwargs)
 
     def head(self, url, **kwargs):
         """
         See :func:`treq.head()`.
         """
+        kwargs.setdefault('_stacklevel', 3)
         return self.request('HEAD', url, **kwargs)
 
     def delete(self, url, **kwargs):
         """
         See :func:`treq.delete()`.
         """
+        kwargs.setdefault('_stacklevel', 3)
         return self.request('DELETE', url, **kwargs)
 
     def request(self, method, url, **kwargs):
@@ -153,15 +150,20 @@ class HTTPClient(object):
         See :func:`treq.request()`.
         """
         method = method.encode('ascii').upper()
+        stacklevel = kwargs.pop('_stacklevel', 2)
 
-        if isinstance(url, unicode):
-            parsed_url = URL.from_text(url)
+        if isinstance(url, DecodedURL):
+            parsed_url = url
+        elif isinstance(url, EncodedURL):
+            parsed_url = DecodedURL(url)
+        elif isinstance(url, six.text_type):
+            parsed_url = DecodedURL.from_text(url)
         else:
-            parsed_url = URL.from_text(url.decode('ascii'))
+            parsed_url = DecodedURL.from_text(url.decode('ascii'))
 
         # Join parameters provided in the URL
         # and the ones passed as argument.
-        params = kwargs.get('params')
+        params = kwargs.pop('params', None)
         if params:
             parsed_url = parsed_url.replace(
                 query=parsed_url.query + tuple(_coerced_query_params(params))
@@ -169,64 +171,18 @@ class HTTPClient(object):
 
         url = parsed_url.to_uri().to_text().encode('ascii')
 
-        # Convert headers dictionary to
-        # twisted raw headers format.
-        headers = kwargs.get('headers')
-        if headers:
-            if isinstance(headers, dict):
-                h = Headers({})
-                for k, v in headers.items():
-                    if isinstance(v, (bytes, unicode)):
-                        h.addRawHeader(k, v)
-                    elif isinstance(v, list):
-                        h.setRawHeaders(k, v)
+        headers = self._request_headers(kwargs.pop('headers', None), stacklevel + 1)
 
-                headers = h
-        else:
-            headers = Headers({})
+        bodyProducer, contentType = self._request_body(
+            data=kwargs.pop('data', None),
+            files=kwargs.pop('files', None),
+            json=kwargs.pop('json', _NOTHING),
+            stacklevel=stacklevel + 1,
+        )
+        if contentType is not None:
+            headers.setRawHeaders(b'Content-Type', [contentType])
 
-        # Here we choose a right producer
-        # based on the parameters passed in.
-        bodyProducer = None
-        data = kwargs.get('data')
-        files = kwargs.get('files')
-        # since json=None needs to be serialized as 'null', we need to
-        # explicitly check kwargs for this key
-        has_json = 'json' in kwargs
-
-        if files:
-            # If the files keyword is present we will issue a
-            # multipart/form-data request as it suits better for cases
-            # with files and/or large objects.
-            files = list(_convert_files(files))
-            boundary = str(uuid.uuid4()).encode('ascii')
-            headers.setRawHeaders(
-                b'content-type', [
-                    b'multipart/form-data; boundary=' + boundary])
-            if data:
-                data = _convert_params(data)
-            else:
-                data = []
-
-            bodyProducer = multipart.MultiPartProducer(
-                data + files, boundary=boundary)
-        elif data:
-            # Otherwise stick to x-www-form-urlencoded format
-            # as it's generally faster for smaller requests.
-            if isinstance(data, (dict, list, tuple)):
-                headers.setRawHeaders(
-                    b'content-type', [b'application/x-www-form-urlencoded'])
-                data = urlencode(data, doseq=True)
-            bodyProducer = self._data_to_body_producer(data)
-        elif has_json:
-            # If data is sent as json, set Content-Type as 'application/json'
-            headers.setRawHeaders(
-                b'content-type', [b'application/json; charset=UTF-8'])
-            content = kwargs['json']
-            json = json_dumps(content, separators=(u',', u':')).encode('utf-8')
-            bodyProducer = self._data_to_body_producer(json)
-
-        cookies = kwargs.get('cookies', {})
+        cookies = kwargs.pop('cookies', {})
 
         if not isinstance(cookies, CookieJar):
             cookies = cookiejar_from_dict(cookies)
@@ -234,8 +190,9 @@ class HTTPClient(object):
         cookies = merge_cookies(self._cookiejar, cookies)
         wrapped_agent = CookieAgent(self._agent, cookies)
 
-        if kwargs.get('allow_redirects', True):
-            if kwargs.get('browser_like_redirects', False):
+        browser_like_redirects = kwargs.pop('browser_like_redirects', False)
+        if kwargs.pop('allow_redirects', True):
+            if browser_like_redirects:
                 wrapped_agent = BrowserLikeRedirectAgent(wrapped_agent)
             else:
                 wrapped_agent = RedirectAgent(wrapped_agent)
@@ -243,7 +200,7 @@ class HTTPClient(object):
         wrapped_agent = ContentDecoderAgent(wrapped_agent,
                                             [(b'gzip', GzipDecoder)])
 
-        auth = kwargs.get('auth')
+        auth = kwargs.pop('auth', None)
         if auth:
             wrapped_agent = add_auth(wrapped_agent, auth)
 
@@ -251,10 +208,12 @@ class HTTPClient(object):
             method, url, headers=headers,
             bodyProducer=bodyProducer)
 
-        timeout = kwargs.get('timeout')
+        reactor = kwargs.pop('reactor', None)
+        if reactor is None:
+            from twisted.internet import reactor
+        timeout = kwargs.pop('timeout', None)
         if timeout:
-            delayedCall = default_reactor(kwargs.get('reactor')).callLater(
-                timeout, d.cancel)
+            delayedCall = reactor.callLater(timeout, d.cancel)
 
             def gotResult(result):
                 if delayedCall.active():
@@ -263,10 +222,139 @@ class HTTPClient(object):
 
             d.addBoth(gotResult)
 
-        if not kwargs.get('unbuffered', False):
+        if not kwargs.pop('unbuffered', False):
             d.addCallback(_BufferedResponse)
 
+        if kwargs:
+            warnings.warn(
+                (
+                    "Got unexpected keyword argument: {}."
+                    " treq will ignore this argument,"
+                    " but will raise TypeError in the next treq release."
+                ).format(", ".join(repr(k) for k in kwargs)),
+                DeprecationWarning,
+                stacklevel=stacklevel,
+            )
+
         return d.addCallback(_Response, cookies)
+
+    def _request_headers(self, headers, stacklevel):
+        """
+        Convert the *headers* argument to a :class:`Headers` instance
+
+        :returns:
+            :class:`twisted.web.http_headers.Headers`
+        """
+        if isinstance(headers, dict):
+            h = Headers({})
+            for k, v in headers.items():
+                if isinstance(v, (bytes, six.text_type)):
+                    h.addRawHeader(k, v)
+                elif isinstance(v, list):
+                    h.setRawHeaders(k, v)
+                else:
+                    warnings.warn(
+                        (
+                            "The value of headers key {!r} has non-string type {}"
+                            " and will be dropped."
+                            " This will raise TypeError in the next treq release."
+                        ).format(k, type(v)),
+                        DeprecationWarning,
+                        stacklevel=stacklevel,
+                    )
+            return h
+        if isinstance(headers, Headers):
+            return headers
+        if headers is None:
+            return Headers({})
+
+        warnings.warn(
+            (
+                "headers must be a dict, twisted.web.http_headers.Headers, or None,"
+                " but found {}, which will be ignored."
+                " This will raise TypeError in the next treq release."
+            ).format(type(headers)),
+            DeprecationWarning,
+            stacklevel=stacklevel,
+        )
+        return Headers({})
+
+    def _request_body(self, data, files, json, stacklevel):
+        """
+        Here we choose a right producer based on the parameters passed in.
+
+        :params data:
+            Arbitrary request body data.
+
+            If *files* is also passed this must be a :class:`dict`,
+            a :class:`tuple` or :class:`list` of field tuples as accepted by
+            :class:`MultiPartProducer`. The request is assigned a Content-Type
+            of ``multipart/form-data``.
+
+            If a :class:`dict`, :class:`list`, or :class:`tuple` it is
+            URL-encoded and the request assigned a Content-Type of
+            ``application/x-www-form-urlencoded``.
+
+            Otherwise, any non-``None`` value is passed to the client's
+            *data_to_body_producer* callable (by default,
+            :class:`IBodyProducer`), which accepts file-like objects.
+
+        :params files:
+            Files to include in the request body, in any of the several formats
+            described in :func:`_convert_files()`.
+
+        :params json:
+            JSON-encodable data, or the sentinel `_NOTHING`. The sentinel is
+            necessary because ``None`` is a valid JSON value.
+        """
+        if json is not _NOTHING and (files or data):
+            warnings.warn(
+                (
+                    "Argument 'json' will be ignored because '{}' was also passed."
+                    " This will raise TypeError in the next treq release."
+                ).format("data" if data else "files"),
+                DeprecationWarning,
+                stacklevel=stacklevel,
+            )
+
+        if files:
+            # If the files keyword is present we will issue a
+            # multipart/form-data request as it suits better for cases
+            # with files and/or large objects.
+            files = list(_convert_files(files))
+            boundary = str(uuid.uuid4()).encode('ascii')
+            if data:
+                data = _convert_params(data)
+            else:
+                data = []
+
+            return (
+                multipart.MultiPartProducer(data + files, boundary=boundary),
+                b'multipart/form-data; boundary=' + boundary,
+            )
+
+        # Otherwise stick to x-www-form-urlencoded format
+        # as it's generally faster for smaller requests.
+        if isinstance(data, (dict, list, tuple)):
+            return (
+                self._data_to_body_producer(urlencode(data, doseq=True)),
+                b'application/x-www-form-urlencoded',
+            )
+        elif data:
+            return (
+                self._data_to_body_producer(data),
+                None,
+            )
+
+        if json is not _NOTHING:
+            return (
+                self._data_to_body_producer(
+                    json_dumps(json, separators=(u',', u':')).encode('utf-8'),
+                ),
+                b'application/json; charset=UTF-8',
+            )
+
+        return None, None
 
 
 def _convert_params(params):
@@ -342,20 +430,20 @@ def _coerced_query_params(params):
     for key, values in items:
         if isinstance(key, bytes):
             key = key.decode('ascii')
-        elif not isinstance(key, unicode):
-            key = unicode(key)
+        elif not isinstance(key, six.text_type):
+            key = six.text_type(key)
         if not isinstance(values, (list, tuple)):
             values = [values]
         for value in values:
             if isinstance(value, bytes):
                 value = value.decode('ascii')
-            elif not isinstance(value, unicode):
-                value = unicode(value)
+            elif not isinstance(value, six.text_type):
+                value = six.text_type(value)
             yield key, value
 
 
 def _from_bytes(orig_bytes):
-    return FileBodyProducer(BytesIO(orig_bytes))
+    return FileBodyProducer(io.BytesIO(orig_bytes))
 
 
 def _from_file(orig_file):
@@ -371,14 +459,12 @@ def _guess_content_type(filename):
 
 
 registerAdapter(_from_bytes, bytes, IBodyProducer)
-registerAdapter(_from_file, BytesIO, IBodyProducer)
+registerAdapter(_from_file, io.BytesIO, IBodyProducer)
 
-if not _PY3:
-    from StringIO import StringIO
-    registerAdapter(_from_file, StringIO, IBodyProducer)
+if six.PY2:
+    registerAdapter(_from_file, six.StringIO, IBodyProducer)
     # Suppress lint failure on Python 3.
     registerAdapter(_from_file, file, IBodyProducer)  # noqa: F821
 else:
-    import io
     # file()/open() equiv on Py3
     registerAdapter(_from_file, io.BufferedReader, IBodyProducer)