+++ src/treq/test/test_multipart.py
@@ -10,16 +10,16 @@ from io import BytesIO
 from twisted.trial import unittest
 from zope.interface.verify import verifyObject
 
-from twisted.python import compat
+from six import PY3, text_type
+
 from twisted.internet import task
 from twisted.web.client import FileBodyProducer
 from twisted.web.iweb import UNKNOWN_LENGTH, IBodyProducer
 
 from treq.multipart import MultiPartProducer, _LengthConsumer
 
-if compat._PY3:
+if PY3:
     long = int
-    unicode = compat.unicode
 
 
 class MultiPartProducerTestCase(unittest.TestCase):
@@ -65,7 +65,7 @@ class MultiPartProducerTestCase(unittest.TestCase):
 
     def newLines(self, value):
 
-        if isinstance(value, unicode):
+        if isinstance(value, text_type):
             return value.replace(u"\n", u"\r\n")
         else:
             return value.replace(b"\n", b"\r\n")